# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api of kindergarden",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tutors": {
            "get": {
                "tags": [
                    "Tutors"
                ],
                "summary": "All tutors",
                "description": "All tutors",
                "operationId": "all_tutor",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "identification": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "parent": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "residence_phone": {
                                            "type": "string",
                                            "format": "string",
                                            "example": ""
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "military": {
                                            "type": "boolean",
                                            "format": "boolean",
                                            "example": ""
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Store tutor",
                "description": "Store tutor",
                "operationId": "store_tutor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Santos"
                                    },
                                    "identification": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "00123492349"
                                    },
                                    "parent": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Tio"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "8293242233"
                                    },
                                    "residence_phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Santo Domingo"
                                    },
                                    "military": {
                                        "type": "boolean",
                                        "format": "boolean",
                                        "example": "true"
                                    },
                                    "kid_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/tutors/{id}": {
            "get": {
                "tags": [
                    "Tutors"
                ],
                "summary": "See tutor",
                "description": "See tutor",
                "operationId": "watch_tutor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "identification": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "parent": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "residence_phone": {
                                            "type": "string",
                                            "format": "string",
                                            "example": ""
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "military": {
                                            "type": "boolean",
                                            "format": "boolean",
                                            "example": "true"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Update tutor",
                "description": "Update tutor",
                "operationId": "update_tutor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "identification": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "parent": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "residence_phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "military": {
                                        "type": "boolean",
                                        "format": "boolean",
                                        "example": "true"
                                    },
                                    "kid_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Delete tutor",
                "description": "Delete tutor",
                "operationId": "delete_tutor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    }
}